requirements.txt
# Hybrid Sparse VAE — Dependencies
torch>=2.1
torchaudio>=2.1
hydra-core>=1.3
omegaconf>=2.3
matplotlib>=3.7
numpy>=1.24
pytest>=7.0
